refactor(fw): Phase 4 COMM module — umbrella, R1.2 -Werror, vendor -isystem
Fifth Phase-4 module (5/7). No folder move (comm/ already a module).

R2.1 — include/comm/comm.h umbrella over the NavLink types, serial channel,
(de)serialiser, iBUS link, and RC buffer. External include sites rewritten
to it (variables.h's comm/channel.h left as a global-coupling exception).

R1.2 / CONV-06 — all comm TUs flipped to full R1.2 + -Werror (13 warnings):
  - empty-brace inits {} -> {0} (channel, serializer); 0b1111 -> 0xF
    (serializer); size/enum->uint8_t and DMA NDTR uint32->uint16 explicit
    casts; (void) on unused params (get_handler_default, flush_task);
    rc_task NDTR cast; imu_telemetry_task / rc_ibus_task got their
    prototypes via the previously-missing vayu_tasks.h include.

Vendor exemption (§5.3): VAYU_R12_WARN_FLAGS now passes vaios/NavHAL
include roots as -isystem, so vendored headers (e.g. rc_task.c pulling
dma_reg.h under -Wconversion) can't fail a vayu module's -Werror build.
This also future-proofs the remaining modules.

Verification: arm target links clean with ACT+EST+SNS+CTRL+COMM under
-Werror; host SITL + full 5-suite ctest green. warning-rollout.md: COMM ✅.

+++ src/comm/rc_task.c
@@ -1,3 +1,4 @@
+#include "vayu_tasks.h"
 #include "comm/ibus.h"
 #include "comm/rc_buffer.h"
 #include "navhal.h"
@@ -90,7 +91,7 @@ void rc_ibus_task(void *args) {
 #endif
     // Current remaining items in circular buffer from DMA NDTR register
     // USART1 is on DMA2, Stream 2
-    uint16_t current_ndtr = DMA2->STREAM[2].NDTR;
+    uint16_t current_ndtr = (uint16_t)DMA2->STREAM[2].NDTR;
 
     // Diagnostic log every 1s (User: Dont remove this this makes the rc not
     // work)
@@ -120,7 +121,7 @@ void rc_ibus_task(void *args) {
         new_data = 1;
         rc_mark_frame_valid();
       }
-      read_ptr = (read_ptr + 1) % IBUS_DMA_BUF_SIZE;
+      read_ptr = (uint16_t)((read_ptr + 1) % IBUS_DMA_BUF_SIZE);
       sys_state_t current_state = system_state_get();
       for (int i = 0; i < 4; i++) {
         if (i != 2 && (ibus_raw_data.channels[i] > 1500 - RADIO_AVOID_BAND &&
